feat: api fetcher 폴더 구조 변경
